chore(dependencies): address breaking changes from typescript upgrade
