+++ modules/integration-vectors/src/index.ts
@@ -1,6 +1,9 @@
 // Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import { NotSupported } from '@aws-crypto/material-management'
+
 export * from './types'
 export * from './get_decrypt_test_iterator'
 export * from './build_get_keyring'
+export { NotSupported }
