+++ modules/material-management/src/needs.ts
@@ -21,3 +21,13 @@ export function needs(
     throw new Err(errorMessage)
   }
 }
+
+export class NotSupported extends Error {
+  code: string
+
+  constructor(message?: string) {
+    super(message)
+    Object.setPrototypeOf(this, NotSupported.prototype)
+    this.code = 'NOT_SUPPORTED'
+  }
+}
